This file is for converting radiance to refl
This assumes you have a panel image inside your tif image. You need to draw a polygon inside the panel and the code will convert the image to reflectance and also give you the factors in case you need to apply this on any other image. Refl = radiance/correction factor